requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
Pygments==2.7.4
Markdown==3.2.1
PyYAML==5.4
django-filter~=2.4.0